requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
dj-database-url==3.1.0
Django==5.2.8
django-allauth==65.13.1
django-cors-headers==4.9.0
django-distill==3.2.7
exceptiongroup==1.3.1
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-genai
googleapis-common-protos
grpcio
grpcio-status
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
jiter==0.12.0
lxml==6.0.2
oauthlib==3.3.1
openai==2.8.1
packaging==25.0
pillow==12.1.0
proto-plus==1.26.1
protobuf
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.2.5
python-decouple==3.8
python-docx==1.2.0
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
websockets==15.0.1
whitenoise==6.11.0
dj3-cloudinary-storage==0.0.6
cloudinary==1.36.0
google-generativeai
